+++ src/finanzplaner/__init__.py
@@ -1,3 +1,3 @@
 """Fin application package."""
 
-__version__ = "1.1.0"
+__version__ = "1.2.0rc1"